Perubahan pada Parameter setiap Class - Long Parameter List
Merubah parameter setiap class menjadi lebih sederhana dan menggunakan  super untuk menghubungkan class utama.

Tujuannya adalah untuk membuat parameter lebih efisien,
